add aggregate.rs
