+++ query/src/sql/planner/binder/select.rs
@@ -20,22 +20,17 @@ use common_ast::ast::Query;
 use common_ast::ast::SelectStmt;
 use common_ast::ast::SetExpr;
 use common_ast::ast::TableReference;
-use common_exception::ErrorCode;
 use common_exception::Result;
 use common_planners::ReadDataSourcePlan;
 use common_planners::SourceInfo;
 
 use crate::sql::optimizer::SExpr;
 use crate::sql::planner::binder::scalar::ScalarBinder;
-use crate::sql::planner::binder::scalar_common::find_aggregate_scalars_from_bind_context;
 use crate::sql::planner::binder::BindContext;
 use crate::sql::planner::binder::Binder;
 use crate::sql::planner::binder::ColumnBinding;
-use crate::sql::planner::binder::ScalarExprRef;
-use crate::sql::plans::AggregatePlan;
 use crate::sql::plans::FilterPlan;
 use crate::sql::plans::LogicalGet;
-use crate::sql::plans::Scalar;
 use crate::sql::IndexType;
 use crate::storages::Table;
 
@@ -75,40 +70,6 @@ impl Binder {
         self.bind_projection(&mut output_context)?;
 
         Ok(output_context)
-    }
-
-    fn analyze_aggregate(
-        &self,
-        output_context: &BindContext,
-        input_context: &mut BindContext,
-    ) -> Result<()> {
-        let mut agg_expr: Vec<ScalarExprRef> = Vec::new();
-        for agg_scalar in find_aggregate_scalars_from_bind_context(output_context)? {
-            match agg_scalar {
-                Scalar::AggregateFunction {
-                    func_name,
-                    distinct,
-                    params,
-                    args,
-                    data_type,
-                    nullable,
-                } => agg_expr.push(Arc::new(Scalar::AggregateFunction {
-                    func_name,
-                    distinct,
-                    params,
-                    args,
-                    data_type,
-                    nullable,
-                })),
-                _ => {
-                    return Err(ErrorCode::LogicalError(
-                        "scalar expr must be Aggregation scalar expr",
-                    ))
-                }
-            }
-        }
-        input_context.agg_scalar_exprs = Some(agg_expr);
-        Ok(())
     }
 
     async fn bind_table_reference(&mut self, stmt: &TableReference) -> Result<BindContext> {
@@ -188,29 +149,6 @@ impl Binder {
         let new_expr =
             SExpr::create_unary(filter_plan.into(), bind_context.expression.clone().unwrap());
         bind_context.expression = Some(new_expr);
-        Ok(())
-    }
-
-    pub(super) fn bind_group_by(
-        &mut self,
-        group_by_expr: &[Expr],
-        input_context: &mut BindContext,
-    ) -> Result<()> {
-        let scalar_binder = ScalarBinder::new();
-        let group_expr = group_by_expr
-            .iter()
-            .map(|expr| scalar_binder.bind_expr(expr, input_context))
-            .collect::<Result<Vec<ScalarExprRef>>>();
-
-        let aggregate_plan = AggregatePlan {
-            group_expr: group_expr?,
-            agg_expr: input_context.agg_scalar_exprs.as_ref().unwrap().clone(),
-        };
-        let new_expr = SExpr::create_unary(
-            aggregate_plan.into(),
-            input_context.expression.clone().unwrap(),
-        );
-        input_context.expression = Some(new_expr);
         Ok(())
     }
 }
